feat: let the users to attach a file from side menu
