+++ src/Screens/SideMenu/Files.tsx
@@ -7,6 +7,7 @@ import {
   FileItemContainer,
   FilesContainer,
   IconsContainer,
+  SideMenuCellAttachNewFile,
   SideMenuCellShowAllFiles,
   SideMenuCellStyled,
   SNIconStyled,
@@ -26,7 +27,7 @@ export const Files: FC<Props> = ({ note }) => {
   const application = useSafeApplicationContext()
 
   const navigation = useNavigation<AppStackNavigationProp<typeof SCREEN_COMPOSE>['navigation']>()
-  const { showActionsMenu, attachedFiles } = useFiles({ note })
+  const { showActionsMenu, handlePressAttachFile, attachedFiles } = useFiles({ note })
 
   const openFilesScreen = () => {
     navigation.navigate(SCREEN_UPLOADED_FILES_LIST, { note })
@@ -60,6 +61,10 @@ export const Files: FC<Props> = ({ note }) => {
           </FileItemContainer>
         )
       })}
+      <SideMenuCellAttachNewFile
+        text={'Attach new file'}
+        onSelect={() => handlePressAttachFile()}
+      />
       <SideMenuCellShowAllFiles
         text={isFilesListTruncated ? 'Show all attached files' : 'Show other files'}
         onSelect={openFilesScreen}
